+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.42.8](https://github.com/i18next/i18next-cli/compare/v1.42.7...v1.42.8) - 2026-02-11
 
 - linter: improve `checkInterpolationParams` (ignore i18next variables) [#178](https://github.com/i18next/i18next-cli/issues/178)
+- extractor: fix ignoreNamespaces not respected when removeUnusedKeys is true [#171](https://github.com/i18next/i18next-cli/issues/171)
 
 ## [1.42.7](https://github.com/i18next/i18next-cli/compare/v1.42.6...v1.42.7) - 2026-02-11